+++ CHANGELOG.md
@@ -19,6 +19,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   ([#825](https://github.com/nix-rust/nix/pull/825))
 - Exposed `MSG_CMSG_CLOEXEC` on *BSD.
   ([#825](https://github.com/nix-rust/nix/pull/825))
+- Added `nix::sys::posix_fallocate` for Android, FreeBSD and Linux.
+  ([#779](https:://github.com/nix-rust/nix/pull/779))
 
 ### Changed
 - Display and Debug for SysControlAddr now includes all fields.